Code cleanup
Performed quite a bit of Code cleanup. 
In particular null checks were added to the methods to prevent errors and remove some duplicate methods.
A method for converting COBRA ids to valid SBML ids was added and integrated when needed in the other methods.
Some unnecessarily complex Code was simplified.
The example model was updated to better reflect the Information given in a COBRAme JSON file.